# api-swagger/api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: '0.0.1'
  title: MusicRoomAPI
host: localhost:3001
basePath: /
schemes:
  # tip: remove http to make production-grade
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  '/api/user/log-in':
    x-swagger-router-controller: login
    post:
      tags:
        - User
      # description: Route to classic log in
      operationId: classicLogin
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/LoginInput'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/LoginResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/api/user/facebook-log-in':
    # binds a127 app logic to a route
    x-swagger-router-controller: login
    post:
      tags:
        - User
      # description: Route to facebook log in
      operationId: facebookLogin
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/SocialLoginInput'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/LoginResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/api/user/google-log-in':
    x-swagger-router-controller: login
    post:
      tags:
        - User
      # description: Route to google log in
      operationId: googleLogin
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/SocialLoginInput'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/LoginResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
...
definitions:
...
  LoginInput:
    required:
      - email
      - password
    properties:
      email:
        type: string
      password:
        type: string
  SocialLoginInput:
    required:
      - email
      - userToken
    properties:
      email:
        type: string
      userToken:
        type: string
  LoginResponse:
    properties:
      sessionId:
        type: string
      user:
        $ref: '#/definitions/User'
  User:
    properties:
      email:
        type: string
      name:
        type: string
      firstname:
        type: string
      avatarUri:
        type: string
      deezer:
        type: boolean
      facebook:
        type: boolean
      google:
        type: boolean
      tags:
        type: array
        items:
          type: string
      privacy:
        type: object
...
  ErrorResponse:
    required:
      - error
    properties:
      error:
        type: string